Refactor formatters into separate functions
